requirements.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
basicsr==1.4.2
beautifulsoup4==4.12.3
certifi==2024.12.14
chainer==7.8.1
char-similar==0.0.1
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
dill==0.3.9
exceptiongroup==1.2.2
facexlib==0.3.0
ffmpeg-python==0.2.0
filelock==3.17.0
filterpy==1.4.5
fonttools==4.55.8
fsspec==2025.2.0
future==1.0.0
gensim==3.7.1
gfpgan==1.3.8
grpcio==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.5.1
lxml==5.3.0
Macropodus==0.0.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
moviepy==2.1.2
mpmath==1.3.0
multiprocess==0.70.17
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddlepaddle==2.6.2
pandas==2.2.3
passlib==1.7.1
pillow==10.4.0
pip==25.0
platformdirs==4.3.6
proglog==0.1.10
protobuf==3.20.2
pyclipper==1.3.0.post6
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
requests==2.32.3
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.8.0
shapely==2.0.6
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tb-nightly==2.19.0a20250205
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.12.12
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.31.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Wand==0.6.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yapf==0.43.0
